[sampling] Feature: Added ability to specify CRS for raster generation in manual txt file.
This is important for reconciling past labels that used specific CRS projections with current method of alphabetically selecting the first CRS. This ensures shapes are the same between downloaded rasters and past labels.

Make sure to keep this commit even if most of the other changes are not retainedfrom refactoring preprocessing.