[preprocess] Feat: Suffix option added for preprocessing
Before the preprocessing scripts only outputted unique directory names
based on size and samples args, but does not allow for different directories
if you change splits, labels or events given. Thus, a suffix option is provided
to distinguish these cases.

Training scripts also will look for a suffix in the data part of the config file
and load in the directory with the suffix added if found.

Ex: if you want to have a texas region only dataset based on the event folders
given, add a suffix='tx'.