[training] WIP: Incorporating NLCD metrics into training scripts
This is only partial part of the NLCD logging mechanism. Certain
design choices still need to be made how to pass these confusion matrices
and NLCD class dicts around, which doesn't work currently with the
wandb logging. So before committing to a design choice a commit
is made here at the fork - this commit is the universal NLCD computation
logic so can always come back to this state.